+++ tests/acceptance/transitions/test_config_band.py
@@ -1,6 +1,7 @@
 """Test ConfigureBand2."""
 
 import pytest
+from ska_mid_dish_dcp_lib.device.b5dc_device_mappings import B5dcFrequency
 
 from ska_mid_dish_manager.models.dish_enums import Band, DishMode
 from tests.utils import remove_subscriptions, setup_subscriptions
@@ -33,7 +34,7 @@ def test_configure_band_a(monitor_tango_servers, event_store_class, dish_manager
 
     [[_], [unique_id]] = dish_manager_proxy.ConfigureBand2(True)
     result_event_store.wait_for_command_result(
-        unique_id, '[0, "SetOperateMode completed"]', timeout=30
+        unique_id, '[0, "SetOperateMode completed."]', timeout=30
     )
     main_event_store.wait_for_value(Band.B2, timeout=30)
     main_event_store.wait_for_value(DishMode.OPERATE, timeout=30)
@@ -79,8 +80,7 @@ def test_configure_band_a(monitor_tango_servers, event_store_class, dish_manager
         ("ConfigureBand3", Band.B3, "B3"),
         ("ConfigureBand4", Band.B4, "B4"),
         ("ConfigureBand5a", Band.B5a, "B5a"),
-        # Special case for 5b until we have a B5DC
-        ("ConfigureBand5b", Band.B1, "B1"),
+        ("ConfigureBand5b", Band.B5b, "B5b"),
         # End up in B2 again
         ("ConfigureBand2", Band.B2, "B2"),
     ],
@@ -115,21 +115,16 @@ def test_configure_band_b(
     main_event_store.wait_for_value(expected_band, timeout=15)
     assert dish_manager_proxy.configuredBand == expected_band
 
-    if band_request == "ConfigureBand5b":
-        expected_progress_updates = [
-            "Fanned out commands: DS.SetIndexPosition, SPFRX.ConfigureBand1",
-            "Awaiting configuredband change to B1",
-            "ConfigureBand1 completed",
-        ]
-    else:
-        expected_progress_updates = [
-            f"Fanned out commands: DS.SetIndexPosition, SPFRX.{band_request}",
-            f"Awaiting configuredband change to {message_str}",
-            f"{band_request} completed",
-        ]
+    spfrx_config_band_cmd = f"SPFRX.{band_request}"
+    if expected_band == Band.B5b:
+        spfrx_config_band_cmd = "SPFRX.ConfigureBand1"
 
+    expected_progress_updates = [
+        f"Fanned out commands: DS.SetIndexPosition, {spfrx_config_band_cmd}",
+        f"Awaiting configuredband change to {message_str}",
+        f"{band_request} complete",
+    ]
     events = status_event_store.wait_for_progress_update(expected_progress_updates[-1], timeout=6)
-
     events_string = "".join([str(event.attr_value.value) for event in events])
 
     # Check that all the expected progress messages appeared
@@ -162,7 +157,7 @@ def test_configure_band_2_from_stow(
     # make sure configuredBand is not B2
     [[_], [unique_id]] = dish_manager_proxy.ConfigureBand1(True)
     result_event_store.wait_for_command_result(
-        unique_id, '[0, "SetOperateMode completed"]', timeout=30
+        unique_id, '[0, "SetOperateMode completed."]', timeout=30
     )
     assert dish_manager_proxy.configuredBand == Band.B1
     assert dish_manager_proxy.dishMode == DishMode.OPERATE
@@ -179,7 +174,7 @@ def test_configure_band_2_from_stow(
 
     [[_], [unique_id]] = dish_manager_proxy.ConfigureBand2(True)
     result_event_store.wait_for_command_result(
-        unique_id, '[0, "ConfigureBand2 completed"]', timeout=30
+        unique_id, '[0, "ConfigureBand2 completed."]', timeout=30
     )
     main_event_store.wait_for_value(Band.B2, timeout=30)
 
@@ -242,7 +237,7 @@ def test_configure_band_json(
     # make sure configuredBand is not B2
     [[_], [unique_id]] = dish_manager_proxy.ConfigureBand(json_payload_1)
     result_event_store.wait_for_command_result(
-        unique_id, '[0, "SetOperateMode completed"]', timeout=30
+        unique_id, '[0, "SetOperateMode completed."]', timeout=30
     )
     assert dish_manager_proxy.configuredBand == Band.B1
     assert dish_manager_proxy.dishMode == DishMode.OPERATE
@@ -265,7 +260,7 @@ def test_configure_band_json(
     """
     [[_], [unique_id]] = dish_manager_proxy.ConfigureBand(json_payload_2)
     result_event_store.wait_for_command_result(
-        unique_id, '[0, "SetOperateMode completed"]', timeout=30
+        unique_id, '[0, "SetOperateMode completed."]', timeout=30
     )
     assert dish_manager_proxy.configuredBand == Band.B2
     assert dish_manager_proxy.dishMode == DishMode.OPERATE
@@ -295,7 +290,7 @@ def test_configure_band_json(
     # but does not set the indexer position again.
     [[_], [unique_id]] = dish_manager_proxy.ConfigureBand(json_payload_2)
     result_event_store.wait_for_command_result(
-        unique_id, '[0, "SetOperateMode completed"]', timeout=30
+        unique_id, '[0, "SetOperateMode completed."]', timeout=30
     )
     assert dish_manager_proxy.configuredBand == Band.B2
     assert dish_manager_proxy.dishMode == DishMode.OPERATE
@@ -314,4 +309,111 @@ def test_configure_band_json(
     for message in expected_progress_updates:
         assert message in events_string
 
+    remove_subscriptions(subscriptions)
+
+
+@pytest.mark.acceptance_incl_b5dc
+def test_configure_band_json_with_b5dc_fanout(
+    monitor_tango_servers,
+    event_store_class,
+    dish_manager_proxy,
+    b5dc_device_proxy,
+):
+    """Test ConfigureBand with receiver band 5b and sub band configuration."""
+    main_event_store = event_store_class()
+    result_event_store = event_store_class()
+    status_event_store = event_store_class()
+
+    dm_attr_cb_mapping = {
+        "dishMode": main_event_store,
+        "configuredBand": main_event_store,
+        "Status": status_event_store,
+        "lrcFinished": result_event_store,
+    }
+
+    # Setup Dish Manager subscriptions
+    subscriptions = setup_subscriptions(dish_manager_proxy, dm_attr_cb_mapping)
+
+    # Setup B5DC proxy subscription to subband attr
+    b5dc_subband_evt_store = event_store_class()
+
+    b5dc_subband_sub = setup_subscriptions(
+        b5dc_device_proxy, {"rfcmFrequency": b5dc_subband_evt_store}
+    )
+
+    # Ensure configuredBand is not B5b
+    json_payload_1 = """
+    {
+        "dish": {
+            "receiver_band": "2",
+            "spfrx_processing_parameters": [
+                {
+                    "dishes": ["all"],
+                    "sync_pps": true
+                }
+            ]
+        }
+    }
+    """
+    [[_], [unique_id]] = dish_manager_proxy.ConfigureBand(json_payload_1)
+    result_event_store.wait_for_finished_command_result(
+        unique_id, "[0, 'SetOperateMode completed.']", timeout=60
+    )
+    assert dish_manager_proxy.configuredBand == Band.B2
+    assert dish_manager_proxy.dishMode == DishMode.OPERATE
+
+    # Ensure B5dc proxy sub band is not 11.1GHz
+    current_sub_band = b5dc_device_proxy.rfcmFrequency
+    if current_sub_band == B5dcFrequency.F_11_1_GHZ.frequency_value_ghz():
+        b5dc_device_proxy.SetFrequency(B5dcFrequency.F_13_2_GHZ)
+
+        # Increased timeout period to account for the polling
+        # loop refreshing the b5dc attribute values
+        b5dc_subband_evt_store.wait_for_value(B5dcFrequency.F_13_2_GHZ.frequency_value_ghz(), 30)
+
+    main_event_store.clear_queue()
+    b5dc_subband_evt_store.clear_queue()
+
+    json_payload_with_sub_band = """
+    {
+        "dish": {
+            "receiver_band": "5b",
+            "sub_band": "1",
+            "spfrx_processing_parameters": [
+                {
+                    "dishes": ["all"],
+                    "sync_pps": true
+                }
+            ]
+        }
+    }
+    """
+    [[_], [unique_id]] = dish_manager_proxy.ConfigureBand(json_payload_with_sub_band)
+    result_event_store.wait_for_finished_command_result(
+        unique_id, "[0, 'SetOperateMode completed.']", timeout=60
+    )
+    assert dish_manager_proxy.configuredBand == Band.B5b
+    assert dish_manager_proxy.dishMode == DishMode.OPERATE
+    assert b5dc_subband_evt_store.wait_for_value(
+        B5dcFrequency.F_11_1_GHZ.frequency_value_ghz(), 30
+    )
+
+    expected_status_updates = [
+        "Fanned out commands: DS.SetIndexPosition, SPFRX.ConfigureBand, B5DC.SetFrequency",
+        "Awaiting DS indexerposition change to B5b",
+        "Awaiting SPFRX configuredband change to B1",
+        "Awaiting B5DC rfcmfrequency change to 11.1",
+        "Awaiting configuredband change to B5b",
+        "DS indexerposition changed to B5b",
+        "SPFRX configuredband changed to B1",
+        "B5DC rfcmfrequency changed to 11.1",
+    ]
+
+    events = status_event_store.get_queue_values(timeout=0)
+
+    events_string = "".join([str(attr_value) for _, attr_value in events])
+    for message in expected_status_updates:
+        assert message in events_string
+
+    remove_subscriptions(b5dc_subband_sub)
     remove_subscriptions(subscriptions)