+++ CHANGELOG.rst
@@ -8,16 +8,42 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 ## unreleased
 *************
 - Verification check added to handle communication status updates when an API_EventTimeout error is received.
+- Updated ConfigureBand commands to automatically apply pointing models upon execution
+- Resolved SKB-1267
+- Added tag user=operator to INFO logs for filtering operator related logs. 
+- Addressed the cleanup of allocated resources when Init() is called.
+
+Version 9.3.0
+*************
+- See RCs for details
+
+Version 9.3.0-rc.5
+******************
+- Fixed WMS device monitoring helm flag logic to correctly enable/disable monitoring.
+
+Version 9.3.0-rc.4
+******************
+- Updated band B5b condition for configureBand command fanned out to SPFRx to check for Band1 (ADR-102).
+- Added new helm flag to enable/disable monitoring of B5DC device and WMS device.
+
+  - `--set dishmanager.b5dcproxy.monitoring=True` (defaults to False)
+  - `--set dishmanager.wms.monitoring=True` (defaults to False)
+
+Version 9.3.0-rc.3
+******************
+- Removed the workaround to cast a string to an int for the band5 downconversion subband field when passing to SPFRx (updated in SPFRx v2.1)
 
 Version 9.3.0-rc.2
 ******************
 - Fix error status node name (errTiltTwoIounit to errTiltTwoIoUnit)
 
 - Upgraded ska-mid-dish-ds-manager to v8.0.3
+
   - Resolve segmentation fault caused by pushing an event with ATTR_INVALID -> ATTR_VALID transition
   - Fix late-binding closure behaviour for error status attributes fget
 
 - Upgraded ska-mid-dish-simulators to v5.6.2.
+
   - Fixed simulator to align the dscstate with the PLC regarding activating/deactiving state.
   - Reduced the servo position frequency updates by a factor of 4.